MPI topology

The MPITopology type defines the MPI Cartesian topology of the decomposition. In other words, it contains information about the number of decomposed dimensions, and the number of processes in each of these dimensions.

Construction

The main MPITopology constructor takes a MPI communicator and a tuple specifying the number of processes in each dimension. For instance, to distribute 12 MPI processes on a $3 × 4$ grid:

comm = MPI.COMM_WORLD  # we assume MPI.Comm_size(comm) == 12
pdims = (3, 4)
topology = MPITopology(comm, pdims)

At the lower level, MPITopology uses MPI_Cart_create to define a Cartesian MPI communicator. For more control, one can also create a Cartesian communicator using MPI.Cart_create, and pass that to MPITopology:

comm = MPI.COMM_WORLD
dims = [3, 4]  # note: array, not tuple!
periods = zeros(Int, N)
reorder = false
comm_cart = MPI.Cart_create(comm, dims, periods, reorder)
topology = MPITopology(comm_cart)

Types

PencilFFTs.Pencils.MPITopologies.MPITopology — Type

MPITopology{N}

Describes an N-dimensional Cartesian MPI decomposition topology.

MPITopology(comm::MPI.Comm, pdims::Dims{N}) where N

Create N-dimensional MPI topology information.

The pdims tuple specifies the number of MPI processes to put in every dimension of the topology. The product of its values must be equal to the number of processes in communicator comm.

Example

# Divide 2D topology into 4×2 blocks.
comm = MPI.COMM_WORLD
@assert MPI.Comm_size(comm) == 8
topology = MPITopology(comm, (4, 2))

MPITopology{N}(comm_cart::MPI.Comm) where N

Create topology information from MPI communicator with Cartesian topology (typically constructed using MPI.Cart_create). The topology must have dimension N.

Methods

PencilFFTs.Pencils.MPITopologies.get_comm — Method

get_comm(t::MPITopology)

Get MPI communicator associated to an MPI Cartesian topology.

Base.length — Method

length(t::MPITopology)

Get total size of Cartesian topology (i.e. total number of MPI processes).

Base.ndims — Method

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

Base.size — Method

size(t::MPITopology)

Get dimensions of Cartesian topology.