Distributed FFT plans

Distributed FFTs are implemented in the PencilFFTs module, and are built on top of PencilArrays.

Creating plans

PencilFFTs.PencilFFTPlan — Type

PencilFFTPlan{T,N}

Plan for N-dimensional FFT-based transform on MPI-distributed data.

PencilFFTPlan(
    size_global::Dims{N}, transforms, proc_dims::Dims{M}, comm::MPI.Comm,
    [real_type = Float64];
    extra_dims = (),
    fftw_flags = FFTW.ESTIMATE,
    fftw_timelimit = FFTW.NO_TIMELIMIT,
    permute_dims = Val(true),
    transpose_method = Transpositions.IsendIrecv(),
    timer = TimerOutput(),
)

Create plan for N-dimensional transform.

Extended help

size_global specifies the global dimensions of the input data.

transforms should be a tuple of length N specifying the transforms to be applied along each dimension. Each element must be a subtype of Transforms.AbstractTransform. For all the possible transforms, see Transform types. Alternatively, transforms may be a single transform that will be automatically expanded into N equivalent transforms. This is illustrated in the example below.

The transforms are applied one dimension at a time, with the leftmost dimension first for forward transforms. For multidimensional FFTs of real data, this means that a real-to-complex FFT must be performed along the first dimension, and then complex-to-complex FFTs are performed along the other two dimensions (see example below).

The data is distributed over the MPI processes in the comm communicator. The distribution is performed over M dimensions (with M < N) according to the values in proc_dims, which specifies the number of MPI processes to put along each dimension.

Optional arguments

The floating point precision can be selected by setting real_type parameter, which is Float64 by default.

extra_dims may be used to specify the sizes of one or more extra dimensions that should not be transformed. These dimensions will be added to the rightmost (i.e. slowest) indices of the arrays. See Extra dimensions below for usage hints.

The keyword arguments fftw_flags and fftw_timelimit are passed to the FFTW plan creation functions (see AbstractFFTs docs).

permute_dims determines whether the indices of the output data should be reversed. For instance, if the input data has global dimensions (Nx, Ny, Nz), then the output of a complex-to-complex FFT would have dimensions (Nz, Ny, Nx). This enables FFTs to always be performed along the first (i.e. fastest) array dimension, which could lead to performance gains. This option is enabled by default. For type inference reasons, it must be a value type (Val(true) or Val(false)).

transpose_method allows to select between implementations of the global data transpositions. See the transpose! docs for details.

timer should be a TimerOutput object. See Measuring performance for details.

Extra dimensions

One possible application of extra_dims is for describing the components of a vector or tensor field. However, this means that different PencilFFTPlans would need to be created for each kind of field (scalar, vector, ...). To avoid the creation of multiple plans, a possibly better alternative is to create tuples (or arrays) of PencilArrays using allocate_input and allocate_output.

Another more legitimate usage of extra_dims is to specify one or more Cartesian dimensions that should not be transformed nor split among MPI processes.

Example

Suppose we want to perform a 3D FFT of real data. The data is to be decomposed along two dimensions, over 8 MPI processes:

size_global = (64, 32, 128)  # size of real input data

# Perform real-to-complex transform along the first dimension, then
# complex-to-complex transforms along the other dimensions.
transforms = (Transforms.RFFT(), Transforms.FFT(), Transforms.FFT())
# transforms = Transforms.RFFT()  # this is equivalent to the above line

proc_dims = (4, 2)  # 2D decomposition
comm = MPI.COMM_WORLD

plan = PencilFFTPlan(size_global, transforms, proc_dims, comm)

Allocating data

PencilFFTs.allocate_input — Function

allocate_input(p::PencilFFTPlan)          -> PencilArray
allocate_input(p::PencilFFTPlan, dims...) -> Array{PencilArray}
allocate_input(p::PencilFFTPlan, Val(N))  -> NTuple{N, PencilArray}

Allocate uninitialised PencilArray that can hold input data for the given plan.

The second and third forms respectively allocate an array of PencilArrays of size dims, and a tuple of NPencilArrays.

In-place plans

If p is an in-place plan, a ManyPencilArray is allocated. This type holds PencilArray wrappers for the input and output transforms (as well as for intermediate transforms) which share the same space in memory. The input and output PencilArrays should be respectively accessed by calling first(::ManyPencilArray) and last(::ManyPencilArray).

Example

Suppose p is an in-place PencilFFTPlan. Then,

@assert is_inplace(p)
A = allocate_input(p) :: ManyPencilArray
v_in = first(A)       :: PencilArray  # input data view
v_out = last(A)       :: PencilArray  # output data view

Also note that in-place plans must be performed directly on the returned ManyPencilArray, and not on the contained PencilArray views:

p * A       # perform forward transform in-place
p \ A       # perform backward transform in-place
# p * v_in  # not allowed!!

PencilFFTs.allocate_output — Function

allocate_output(p::PencilFFTPlan)          -> PencilArray
allocate_output(p::PencilFFTPlan, dims...) -> Array{PencilArray}
allocate_output(p::PencilFFTPlan, Val(N))  -> NTuple{N, PencilArray}

Allocate uninitialised PencilArray that can hold output data for the given plan.

If p is an in-place plan, a ManyPencilArray is allocated.

See allocate_input for details.

Methods

PencilFFTs.Pencils.MPITopologies.get_comm — Method

get_comm(p::PencilFFTPlan)

Get MPI communicator associated to a PencilFFTPlan.

PencilFFTs.get_scale_factor — Method

get_scale_factor(p::PencilFFTPlan)

Get scale factor associated to a PencilFFTPlan.

PencilFFTs.Pencils.get_timer — Method

get_timer(p::PencilFFTPlan)

Get TimerOutput attached to a PencilFFTPlan.

See Measuring performance for details.

PencilFFTs.Transforms.is_inplace — Method

Transforms.is_inplace(p::PencilFFTPlan)

Returns true if the given plan operates in-place on the input data, false otherwise.