Pencil configurations

A pencil configuration refers to a given distribution of multidimensional data among MPI processes. This information is encoded in the Pencil type.

A pencil configuration includes:

MPI topology information,

global and local dimensions of the numerical grid,

subset of decomposed dimensions,

type of decomposed data (e.g. Float64),

definition of optional permutation of dimensions.

PencilFFTs.Pencils — Module

Pencils

Base module defining types for multidimensional data decompositions using MPI.

Construction

The creation of a new Pencil requires a MPITopology, as well as the global data dimensions and a list of decomposed dimensions. Optionally, one can also specify the data type (the default is Float64) and a permutation of dimensions.

For instance, to decompose along the first and third dimensions of a complex 3D dataset,

topology = MPITopology(#= ... =#)
dims_global = (16, 32, 64)
decomp_dims = (1, 3)  # this requires ndims(topology) == 2
pencil = Pencil(topology, dims_global, decomp_dims, Complex{Float64})

One may also want to create multiple pencil configurations that differ, for instance, on the selection of decomposed dimensions. For this case, a second constructor is available that takes an already existing Pencil instance. Calling this constructor should be preferred when possible since it allows sharing memory buffers (used for instance for global transpositions) and thus reducing memory usage. The following creates a Pencil equivalent to the one above, but with different decomposed dimensions:

pencil_x = Pencil(pencil, decomp_dims=(2, 3))

See the Pencil documentation for more details.

Dimension permutations

As mentioned above, a Pencil may optionally be given information on dimension permutations. In this case, the layout of the data arrays in memory is different from the logical order of dimensions.

To make this clearer, consider the example above where the global data dimensions are $N_x × N_y × N_z = 16 × 32 × 64$. In this case, the logical order is $(x, y, z)$. Now let's say that we want the memory order of the data to be $(y, z, x)$,[1] which corresponds to the permutation (2, 3, 1).

Permutations are passed to the Pencil constructor via the permute keyword argument. For performance reasons, dimension permutations are compile-time constants, and they should be specified using the Permutation type defined in PencilArrays. For instance,

permutation = Permutation(2, 3, 1)
pencil = Pencil(#= ... =#, permute=permutation)

One can also pass NoPermutation as a permutation, which disables permutations (this is the default).

Types

PencilFFTs.Pencils.Pencil — Type

Pencil{N,M,T}

Describes the decomposition of an N-dimensional Cartesian geometry among MPI processes along M directions (with M < N).

The Pencil describes the decomposition of arrays of element type T.

Pencil(topology::MPITopology{M}, size_global::Dims{N},
       decomp_dims::Dims{M}, [element_type=Float64];
       permute::Permutation=NoPermutation(),
       timer=TimerOutput())

Define the decomposition of an N-dimensional geometry along M dimensions.

The dimensions of the geometry are given by size_global = (N1, N2, ...). The Pencil describes the decomposition of an array of dimensions size_global and type T across a group of MPI processes.

Data is distributed over the given M-dimensional MPI topology (with M < N). The decomposed dimensions are given by decomp_dims.

The optional parameter perm should be a (compile-time) tuple defining a permutation of the data indices. Such permutation may be useful for performance reasons, since it may be preferable (e.g. for FFTs) that the data is contiguous along the pencil direction.

It is also possible to pass a TimerOutput to the constructor. See Measuring performance for details.

Examples

Decompose a 3D geometry of global dimensions $N_x × N_y × N_z = 4×8×12$ along the second ($y$) and third ($z$) dimensions.

Pencil(topology, (4, 8, 12), (2, 3))                          # data is in (x, y, z) order
Pencil(topology, (4, 8, 12), (2, 3), permute=Permutation(3, 2, 1))  # data is in (z, y, x) order

In the second case, the actual data is stored in (z, y, x) order within each MPI process.

Pencil(p::Pencil{N,M}, [element_type=eltype(p)];
       decomp_dims::Dims{M}=get_decomposition(p),
       size_global::Dims{N}=size_global(p),
       permute::P=get_permutation(p),
       timer::TimerOutput=get_timer(p))

Create new pencil configuration from an existent one.

This constructor enables sharing temporary data buffers between the two pencil configurations, leading to reduced global memory usage.

PencilFFTs.Permutations.Permutation — Type

Permutation{p}

Describes a compile-time dimension permutation.

The type parameter p should be a valid permutation such as (3, 1, 2).

The parameter p may also be nothing. A Permutation{nothing} represents an identity permutation: it is equivalent to Permutation{(1, 2, …, N)} for any number of dimensions N. The NoPermutation type is provided as an alias for Permutation{nothing}.

Permutation(perm::Vararg{Int})
Permutation(perm::NTuple{N,Int})

Constructs a Permutation.

Example

Both are equivalent:

p1 = Permutation(3, 4)
p2 = Permutation((3, 4))

Permutation(nothing)
NoPermutation()

Constructs an identity permutation.

PencilFFTs.Permutations.NoPermutation — Type

NoPermutation

Alias for an identity permutation, i.e. NoPermutation = Permutation{nothing}.

This alias can be called as a constructor: NoPermutation() creates a Permutation{nothing}, as expected.

Methods

Base.eltype — Method

eltype(Pencil)
eltype(p::Pencil)

Element type associated to the given pencil type.

PencilFFTs.Pencils.MPITopologies.get_comm — Method

get_comm(p::Pencil)

Get MPI communicator associated to an MPI decomposition scheme.

PencilFFTs.Pencils.get_decomposition — Method

get_decomposition(p::Pencil)

Get tuple with decomposed dimensions of the given pencil configuration.

PencilFFTs.Pencils.get_permutation — Method

get_permutation(p::Pencil)

Get index permutation associated to the given pencil configuration.

Returns NoPermutation() if there is no associated permutation.

Base.length — Method

length(p::Pencil)

Get linear length of data associated to the local pencil layout.

Base.ndims — Method

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

PencilFFTs.Pencils.range_local — Method

range_local(p::Pencil; permute=false)

Local data range held by the pencil.

By default the dimensions are not permuted, i.e. they follow the logical order of dimensions.

PencilFFTs.Pencils.size_global — Method

size_global(p::Pencil; permute=false)

Global dimensions of the Cartesian grid associated to the given domain decomposition.

Like size_local, by default the returned dimensions are not permuted.

PencilFFTs.Pencils.size_local — Method

size_local(p::Pencil; permute=false)

Local dimensions of the data held by the pencil.

By default the dimensions are not permuted, i.e. they follow the logical order of dimensions.

PencilFFTs.Pencils.to_local — Method

to_local(p::Pencil, global_inds; permute=false)

Convert non-permuted global indices to local indices.

Indices can be optionally permuted using the permutation associated to the pencil configuration p.

1Why would we want this? Perhaps because we want to efficiently perform FFTs along $y$, which, under this permutation, would be the fastest dimension.