Permutation tools

Tools for working with Permutation objects.

PencilFFTs.Permutations.append_to_permutation — Function

append_to_permutation(p::Permutation, ::Val{M})

Append M non-permuted dimensions to the given permutation.

Examples

julia> append_to_permutation(Permutation(2, 3, 1), Val(2))
Permutation(2, 3, 1, 4, 5)

julia> append_to_permutation(NoPermutation(), Val(2))
NoPermutation()

PencilFFTs.Permutations.check_permutation — Function

check_permutation(perm)

Check the validity of a Permutation.

Throws ArgumentError if the permutation is invalid.

Examples

julia> check_permutation(Permutation(3, 2, 1))  # no error

julia> check_permutation(NoPermutation())       # no error

julia> check_permutation(Permutation(3, 3, 1))
ERROR: ArgumentError: invalid permutation of dimensions: Permutation(3, 3, 1)

PencilFFTs.Permutations.inverse_permutation — Function

inverse_permutation(p::Permutation)

Returns the inverse permutation of p.

Functionally equivalent to Julia's invperm, with the advantage that the result is a compile time constant.

See also relative_permutation.

Examples

julia> p = Permutation(2, 3, 1);

julia> q = inverse_permutation(p)
Permutation(3, 1, 2)

julia> t_orig = (36, 42, 14);

julia> t_perm = permute_indices(t_orig, p)
(42, 14, 36)

julia> permute_indices(t_perm, q) === t_orig
true

PencilFFTs.Permutations.is_identity_permutation — Function

is_identity_permutation(p::Permutation)

Returns true if p is an identity permutation, i.e. if it is equivalent to (1, 2, 3, ...).

julia> is_identity_permutation(Permutation(1, 2, 3))
true

julia> is_identity_permutation(Permutation(1, 3, 2))
false

julia> is_identity_permutation(NoPermutation())
true

PencilFFTs.Permutations.permute_indices — Function

permute_indices(indices, perm::Permutation)

Permute indices according to a compile-time permutation.

indices may be a Tuple of indices, a CartesianIndex, or a Permutation to be reordered according to perm.

Examples

julia> perm = Permutation(2, 3, 1);

julia> permute_indices((36, 42, 14), perm)
(42, 14, 36)

julia> permute_indices(CartesianIndex(36, 42, 14), perm)
CartesianIndex(42, 14, 36)

julia> permute_indices(Permutation(3, 1, 2), perm)
Permutation(1, 2, 3)

PencilFFTs.Permutations.relative_permutation — Function

relative_permutation(x::Permutation, y::Permutation)

Get relative permutation needed to get from x to y. That is, the permutation perm such that permute_indices(x, perm) == y.

The computation is performed at compile time using generated functions.

Examples

julia> x = Permutation(3, 1, 2);

julia> y = Permutation(2, 1, 3);

julia> perm = relative_permutation(x, y)
Permutation(3, 2, 1)

julia> permute_indices(x, perm) == y
true