Array wrappers

The PencilArrays module defines types for handling MPI-distributed data.

The most important types are:

PencilArray: array wrapper including MPI decomposition information. Takes local indices starting at 1, regardless of the location of each MPI process on the global topology.

GlobalPencilArray: PencilArray wrapper that takes global indices, which generally don't start at 1. See also Global views.

PencilFFTs.PencilArrays — Module

PencilArrays

Array wrappers for MPI-distributed data.

Construction

A PencilArray wrapper can be constructed from a Pencil instance as

pencil = Pencil(#= ... =#)
A = PencilArray(pencil)
parent(A)  # returns the allocated Array

This allocates a new Array with the local dimensions and data type associated to the Pencil.

One can also construct a PencilArray wrapper from an existing AbstractArray, whose dimensions and type must be compatible with the Pencil configuration. For instance, the following works:

T = eltype(pencil)
dims = size_local(pencil, permute=true)  # dimensions of data array must be permuted!
data = Array{T}(undef, dims)
A = PencilArray(pencil, data)

Note that data does not need to be a Array, but can be any subtype of AbstractArray.

Dimension permutations

Unlike the wrapped AbstractArray, the PencilArray wrapper takes non-permuted indices. For instance, if the underlying permutation of the Pencil is (2, 3, 1), then A[i, j, k] points to the same value as parent(A)[j, k, i].

Global views

PencilArrays are accessed using local indices that start at 1, regardless of the location of the subdomain associated to the local process on the global grid. Sometimes it may be more convenient to use global indices describing the position of the local process in the domain. For this, the global_view function is provided that generates an OffsetArray wrapper taking global indices. For more details, see for instance the gradient example.

Types

PencilFFTs.PencilArrays.PencilArray — Type

PencilArray(pencil::Pencil, data::AbstractArray{T,N})

Create array wrapper with pencil decomposition information.

The array dimensions and element type must be consistent with those of the given pencil.

Index permutations

If the Pencil has an associated index permutation, then data must have its dimensions permuted accordingly.

Unlike data, the resulting PencilArray should be accessed with unpermuted indices.

Example

Suppose pencil has local dimensions (10, 20, 30) before permutation, and has an asociated permutation (2, 3, 1). Then:

data = zeros(20, 30, 10)       # parent array (with permuted dimensions)

u = PencilArray(pencil, data)  # wrapper with dimensions (10, 20, 30)
@assert size(u) === (10, 20, 30)

u[15, 25, 5]          # BoundsError (15 > 10 and 25 > 20)
u[5, 15, 25]          # correct
parent(u)[15, 25, 5]  # correct

Extra dimensions

The data array can have one or more extra dimensions to the right (slow indices), which are not affected by index permutations.

Example

dims = (20, 30, 10)
PencilArray(pencil, zeros(dims...))        # works (scalar)
PencilArray(pencil, zeros(dims..., 3))     # works (3-component vector)
PencilArray(pencil, zeros(dims..., 4, 3))  # works (4×3 tensor)
PencilArray(pencil, zeros(3, dims...))     # fails

PencilArray(pencil::Pencil, [extra_dims=()])

Allocate an uninitialised PencilArray that can hold data in the local pencil.

Extra dimensions, for instance representing vector components, can be specified. These dimensions are added to the rightmost (slowest) indices of the resulting array.

Example

Suppose pencil has local dimensions (20, 10, 30) after permutation. Then:

PencilArray(pencil)          # array dimensions are (20, 10, 30)
PencilArray(pencil, (4, 3))  # array dimensions are (20, 10, 30, 4, 3)

PencilFFTs.PencilArrays.GlobalPencilArray — Type

GlobalPencilArray{T,N} <: AbstractArray{T,N}

Alias for an OffsetArray wrapping a PencilArray.

Unlike PencilArrays, GlobalPencilArrays take global indices, which in general don't start at 1 for a given MPI process.

The global_view function should be used to create a GlobalPencilArray from a PencilArray.

PencilFFTs.PencilArrays.PencilArrayCollection — Constant

PencilArrayCollection

UnionAll type describing a collection of PencilArrays.

Such a collection can be a tuple or an array of PencilArrays.

Collections are by assumption homogeneous: each array has the same properties, and in particular, is associated to the same Pencil configuration.

For convenience, certain operations defined for PencilArray are also defined for PencilArrayCollection, and return the same value as for a single PencilArray. Some examples are pencil, range_local and get_comm.

Also note that functions from Base, such as size or ndims, are not overloaded for PencilArrayCollection, since they already have a definition for tuples and arrays (and redefining them would be type piracy...).

PencilFFTs.PencilArrays.MaybePencilArrayCollection — Constant

MaybePencilArrayCollection

UnionAll type representing either a PencilArray or a collection of PencilArrays.

See also PencilArrayCollection.

PencilFFTs.PencilArrays.ManyPencilArray — Type

ManyPencilArray{T,N,M}

Container holding M different PencilArray views to the same underlying data buffer. All views share the same element type T and dimensionality N.

This can be useful to perform in-place operations on PencilArray data.

ManyPencilArray(pencils...; extra_dims=())

Create a ManyPencilArray container that can hold data associated to all the given Pencils.

The optional extra_dims argument is the same as for PencilArray.

Methods

PencilArray

PencilFFTs.PencilArrays.extra_dims — Method

extra_dims(x::PencilArray)
extra_dims(x::PencilArrayCollection)

Return tuple with size of "extra" dimensions of PencilArray.

PencilFFTs.Pencils.MPITopologies.get_comm — Method

get_comm(x::PencilArray)
get_comm(x::PencilArrayCollection)

Get MPI communicator associated to a pencil-distributed array.

PencilFFTs.Pencils.get_permutation — Method

get_permutation(x::PencilArray)
get_permutation(x::PencilArrayCollection)

Get index permutation associated to the given PencilArray.

Returns NoPermutation() if there is no associated permutation.

PencilFFTs.PencilArrays.global_view — Method

global_view(x::PencilArray)

Create an OffsetArray of a PencilArray that takes unpermuted global indices.

PencilFFTs.PencilArrays.ndims_extra — Method

ndims_extra(x::PencilArray)
ndims_extra(x::PencilArrayCollection)

Number of "extra" dimensions associated to PencilArray.

These are the dimensions that are not associated to the domain geometry. For instance, they may correspond to vector or tensor components.

These dimensions correspond to the rightmost indices of the array.

The total number of dimensions of a PencilArray is given by:

ndims(x) == ndims_space(x) + ndims_extra(x)

PencilFFTs.PencilArrays.ndims_space — Method

ndims_space(x::PencilArray)
ndims_space(x::PencilArrayCollection)

Number of dimensions associated to the domain geometry.

These dimensions correspond to the leftmost indices of the array.

The total number of dimensions of a PencilArray is given by:

ndims(x) == ndims_space(x) + ndims_extra(x)

Base.parent — Method

parent(x::PencilArray)

Return array wrapped by a PencilArray.

PencilFFTs.PencilArrays.pencil — Method

pencil(x::PencilArray)

Return decomposition configuration associated to a PencilArray.

Base.pointer — Method

pointer(x::PencilArray)

Return pointer to the start of the underlying data.

Use with caution: this may not make a lot of sense if the underlying data is not contiguous or strided (e.g. if the PencilArray is wrapping a non-strided SubArray).

PencilFFTs.Pencils.range_local — Method

range_local(x::PencilArray; permute=false)
range_local(x::PencilArrayCollection; permute=false)

Local data range held by the PencilArray.

By default the dimensions are not permuted, i.e. they are returned in logical order.

Base.size — Method

size(x::PencilArray)

Return logical (unpermuted) local dimensions of a PencilArray.

PencilFFTs.Pencils.size_local — Method

size_local(x::PencilArray; permute=false)
size_local(x::PencilArrayCollection; permute=false)

Local dimensions of the data held by the PencilArray.

If permute=false, this is the same as size(x).

See also size_local(::Pencil).

PencilFFTs.Pencils.size_global — Method

size_global(x::PencilArray; permute=false)
size_global(x::PencilArrayCollection; permute=false)

Global dimensions associated to the given array.

By default, the logical (unpermuted) dimensions of the dataset are returned.

See also size_global(::Pencil).

ManyPencilArray

Base.first — Method

first(A::ManyPencilArray)

Returns the first PencilArray wrapped by A.

Base.getindex — Method

getindex(A::ManyPencilArray, ::Val{i})
getindex(A::ManyPencilArray, i::Integer)

Returns the i-th PencilArray wrapped by A.

If possible, the Val{i} form should be preferred, as it is more efficient and it allows the compiler to know the return type.

See also first(::ManyPencilArray), last(::ManyPencilArray).

Example

A = ManyPencilArray(pencil1, pencil2, pencil3)

# Get the PencilArray associated to `pencil2`.
# u2 = A[2]
u2 = A[Val(2)]  # faster!

Base.last — Method

last(A::ManyPencilArray)

Returns the last PencilArray wrapped by A.

Base.length — Method

length(A::ManyPencilArray)

Returns the number of PencilArrays wrapped by A.