MPI-distributed data

The distribution of global data among MPI processes is managed by the PencilArrays module. This module may be used independently of the FFT functionality.

The PencilArrays module defines types that describe an MPI Cartesian topology and the decomposition of data over MPI processes. The module also defines array wrappers, most notably the PencilArray type, which allow to conveniently and efficiently work with MPI-decomposed data.

PencilFFTs.PencilArrays — Module

Module for multidimensional data decomposition using MPI.

Handles different decomposition configurations and data transpositions between them. Also defines relevant data structures for handling distributed data.

MPI topology

The MPITopology type defines the MPI Cartesian topology of the decomposition. In other words, it contains information about the number of decomposed dimensions, and the number of processes in each of these dimensions.

Construction

The main MPITopology constructor takes a MPI communicator and a tuple specifying the number of processes in each dimension. For instance, to distribute 12 MPI processes on a $3 × 4$ grid:

comm = MPI.COMM_WORLD  # we assume MPI.Comm_size(comm) == 12
pdims = (3, 4)
topology = MPITopology(comm, pdims)

At the lower level, MPITopology uses MPI_Cart_create to define a Cartesian MPI communicator. For more control, one can also create a Cartesian communicator using MPI.Cart_create, and pass that to MPITopology:

comm = MPI.COMM_WORLD
dims = [3, 4]  # note: array, not tuple!
periods = zeros(Int, N)
reorder = false
comm_cart = MPI.Cart_create(comm, dims, periods, reorder)
topology = MPITopology(comm_cart)

Types

PencilFFTs.PencilArrays.MPITopologies.MPITopology — Type

MPITopology{N}

Describes an N-dimensional Cartesian MPI decomposition topology.

MPITopology(comm::MPI.Comm, pdims::Dims{N}) where N

Create N-dimensional MPI topology information.

The pdims tuple specifies the number of MPI processes to put in every dimension of the topology. The product of its values must be equal to the number of processes in communicator comm.

Example

# Divide 2D topology into 4×2 blocks.
comm = MPI.COMM_WORLD
@assert MPI.Comm_size(comm) == 8
topology = MPITopology(comm, (4, 2))

MPITopology{N}(comm_cart::MPI.Comm) where N

Create topology information from MPI communicator with Cartesian topology (typically constructed using MPI.Cart_create). The topology must have dimension N.

Methods

PencilFFTs.PencilArrays.MPITopologies.get_comm — Method

get_comm(t::MPITopology)

Get MPI communicator associated to an MPI Cartesian topology.

Base.length — Method

length(t::MPITopology)

Get total size of Cartesian topology (i.e. total number of MPI processes).

Base.ndims — Method

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

Base.size — Method

size(t::MPITopology)

Get dimensions of Cartesian topology.

Pencil configurations

A pencil configuration refers to a given distribution of multidimensional data among MPI processes. This information is encoded in the Pencil type.

More precisely, a pencil configuration includes:

MPI topology information,

global and local dimensions of the numerical grid,

subset of decomposed dimensions,

type of decomposed data (e.g. Float64),

definition of optional permutation of dimensions.

Construction

The creation of a new Pencil requires a MPITopology, as well as the global data dimensions and a list of decomposed dimensions. Optionally, one can also specify the data type (the default is Float64) and a permutation of dimensions.

For instance, to decompose along the first and third dimensions of a complex 3D dataset,

topology = MPITopology(#= ... =#)
dims_global = (16, 32, 64)
decomp_dims = (1, 3)  # this requires ndims(topology) == 2
pencil = Pencil(topology, dims_global, decomp_dims, Complex{Float64})

One may also want to create multiple pencil configurations that differ, for instance, on the selection of decomposed dimensions. For this case, a second constructor is available that takes an already existing Pencil instance. Calling this constructor should be preferred when possible since it allows sharing memory buffers (used for instance for global transpositions) and thus reducing memory usage. The following creates a Pencil equivalent to the one above, but with different decomposed dimensions:

pencil_x = Pencil(pencil, decomp_dims=(2, 3))

See the Pencil documentation for more details.

Dimension permutations

As mentioned above, a Pencil may optionally be given information on dimension permutations. In this case, the layout of the data arrays in memory is different from the logical order of dimensions.

To make this clearer, consider the example above where the global data dimensions are $N_x × N_y × N_z = 16 × 32 × 64$. In this case, the logical order is $(x, y, z)$. Now let's say that we want the memory order of the data to be $(y, z, x)$,[1] which corresponds to the permutation (2, 3, 1).

Permutations are passed to the Pencil constructor via the permute keyword argument. For performance reasons, dimension permutations are compile-time constants, and they should be specified using the Permutation type defined in PencilArrays. For instance,

permutation = Permutation(2, 3, 1)
pencil = Pencil(#= ... =#, permute=permutation)

One can also pass NoPermutation as a permutation, which disables permutations (this is the default).

Types

PencilFFTs.PencilArrays.Pencil — Type

Pencil{N,M,T}

Describes the decomposition of an N-dimensional Cartesian geometry among MPI processes along M directions (with M < N).

The Pencil describes the decomposition of arrays of element type T.

Pencil(topology::MPITopology{M}, size_global::Dims{N},
       decomp_dims::Dims{M}, [element_type=Float64];
       permute::Permutation=NoPermutation(),
       timer=TimerOutput())

Define the decomposition of an N-dimensional geometry along M dimensions.

The dimensions of the geometry are given by size_global = (N1, N2, ...). The Pencil describes the decomposition of an array of dimensions size_global and type T across a group of MPI processes.

Data is distributed over the given M-dimensional MPI topology (with M < N). The decomposed dimensions are given by decomp_dims.

The optional parameter perm should be a (compile-time) tuple defining a permutation of the data indices. Such permutation may be useful for performance reasons, since it may be preferable (e.g. for FFTs) that the data is contiguous along the pencil direction.

It is also possible to pass a TimerOutput to the constructor. See Measuring performance for details.

Examples

Decompose a 3D geometry of global dimensions $N_x × N_y × N_z = 4×8×12$ along the second ($y$) and third ($z$) dimensions.

Pencil(topology, (4, 8, 12), (2, 3))                          # data is in (x, y, z) order
Pencil(topology, (4, 8, 12), (2, 3), permute=Permutation(3, 2, 1))  # data is in (z, y, x) order

In the second case, the actual data is stored in (z, y, x) order within each MPI process.

Pencil(p::Pencil{N,M}, [element_type=eltype(p)];
       decomp_dims::Dims{M}=get_decomposition(p),
       size_global::Dims{N}=size_global(p),
       permute::P=get_permutation(p),
       timer::TimerOutput=get_timer(p))

Create new pencil configuration from an existent one.

This constructor enables sharing temporary data buffers between the two pencil configurations, leading to reduced global memory usage.

PencilFFTs.PencilArrays.Permutation — Type

Permutation{p}

Describes a compile-time dimension permutation.

The type parameter p should be a valid permutation such as (3, 1, 2).

The parameter p may also be nothing. A Permutation{nothing} represents an identity permutation: it is equivalent to Permutation{(1, 2, …, N)} for any number of dimensions N. The NoPermutation type is provided as an alias for Permutation{nothing}.

Permutation(perm::Vararg{Int})
Permutation(perm::NTuple{N,Int})

Constructs a Permutation.

Example

Both are equivalent:

p1 = Permutation(3, 4)
p2 = Permutation((3, 4))

Permutation(nothing)
NoPermutation()

Constructs an identity permutation.

PencilFFTs.PencilArrays.NoPermutation — Type

NoPermutation

Alias for an identity permutation, i.e. NoPermutation = Permutation{nothing}.

This alias can be called as a constructor: NoPermutation() creates a Permutation{nothing}, as expected.

Methods

Base.eltype — Method

eltype(Pencil)
eltype(p::Pencil)

Element type associated to the given pencil type.

PencilFFTs.PencilArrays.MPITopologies.get_comm — Method

get_comm(p::Pencil)

Get MPI communicator associated to an MPI decomposition scheme.

PencilFFTs.PencilArrays.get_decomposition — Method

get_decomposition(p::Pencil)

Get tuple with decomposed dimensions of the given pencil configuration.

PencilFFTs.PencilArrays.get_permutation — Method

get_permutation(p::Pencil)

Get index permutation associated to the given pencil configuration.

Returns NoPermutation() if there is no associated permutation.

Base.length — Method

length(p::Pencil)

Get linear length of data associated to the local pencil layout.

Base.ndims — Method

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

PencilFFTs.PencilArrays.range_local — Method

range_local(p::Pencil; permute=false)

Local data range held by the pencil.

By default the dimensions are not permuted, i.e. they follow the logical order of dimensions.

PencilFFTs.PencilArrays.size_global — Method

size_global(p::Pencil; permute=false)

Global dimensions of the Cartesian grid associated to the given domain decomposition.

Like size_local, by default the returned dimensions are not permuted.

PencilFFTs.PencilArrays.size_local — Method

size_local(p::Pencil; permute=false)

Local dimensions of the data held by the pencil.

By default the dimensions are not permuted, i.e. they follow the logical order of dimensions.

PencilFFTs.PencilArrays.to_local — Method

to_local(p::Pencil, global_inds; permute=false)

Convert non-permuted global indices to local indices.

Indices can be optionally permuted using the permutation associated to the pencil configuration p.

Array wrappers

The PencilArrays module also defines a PencilArray type that wraps an AbstractArray while including pencil decomposition information.

Construction

A PencilArray wrapper can be simply constructed from a Pencil instance as

pencil = Pencil(#= ... =#)
A = PencilArray(pencil)
parent(A)  # returns the allocated Array

This allocates a new Array with the local dimensions and data type associated to the Pencil.

One can also construct a PencilArray wrapper from an existing AbstractArray, whose dimensions and type must be compatible with the Pencil configuration. For instance, the following works:

T = eltype(pencil)
dims = size_local(pencil, permute=true)  # dimensions of data array must be permuted!
data = Array{T}(undef, dims)
A = PencilArray(pencil, data)

Note that data does not need to be a Array, but can be any subtype of AbstractArray.

Dimension permutations

Unlike the wrapped AbstractArray, the PencilArray wrapper takes non-permuted indices. For instance, if the underlying permutation of the Pencil is (2, 3, 1), then A[i, j, k] points to the same value as parent(A)[j, k, i].

Global views

PencilArrays are accessed using local indices that start at 1, regardless of the location of the subdomain associated to the local process on the global grid. Sometimes it may be more convenient to use global indices describing the position of the local process in the domain. For this, the global_view function is provided that generates an OffsetArray wrapper taking global indices. For more details, see for instance the gradient example.

Types

PencilFFTs.PencilArrays.PencilArray — Type

PencilArray(pencil::Pencil, data::AbstractArray{T,N})

Create array wrapper with pencil decomposition information.

The array dimensions and element type must be consistent with those of the given pencil.

Index permutations

If the Pencil has an associated index permutation, then data must have its dimensions permuted accordingly.

Unlike data, the resulting PencilArray should be accessed with unpermuted indices.

Example

Suppose pencil has local dimensions (10, 20, 30) before permutation, and has an asociated permutation (2, 3, 1). Then:

data = zeros(20, 30, 10)       # parent array (with permuted dimensions)

u = PencilArray(pencil, data)  # wrapper with dimensions (10, 20, 30)
@assert size(u) === (10, 20, 30)

u[15, 25, 5]          # BoundsError (15 > 10 and 25 > 20)
u[5, 15, 25]          # correct
parent(u)[15, 25, 5]  # correct

Extra dimensions

The data array can have one or more extra dimensions to the right (slow indices), which are not affected by index permutations.

Example

dims = (20, 30, 10)
PencilArray(pencil, zeros(dims...))        # works (scalar)
PencilArray(pencil, zeros(dims..., 3))     # works (3-component vector)
PencilArray(pencil, zeros(dims..., 4, 3))  # works (4×3 tensor)
PencilArray(pencil, zeros(3, dims...))     # fails

PencilArray(pencil::Pencil, [extra_dims=()])

Allocate an uninitialised PencilArray that can hold data in the local pencil.

Extra dimensions, for instance representing vector components, can be specified. These dimensions are added to the rightmost (slowest) indices of the resulting array.

Example

Suppose pencil has local dimensions (20, 10, 30) after permutation. Then:

PencilArray(pencil)          # array dimensions are (20, 10, 30)
PencilArray(pencil, (4, 3))  # array dimensions are (20, 10, 30, 4, 3)

PencilFFTs.PencilArrays.PencilArrayCollection — Constant

PencilArrayCollection

UnionAll type describing a collection of PencilArrays.

Such a collection can be a tuple or an array of PencilArrays.

Collections are by assumption homogeneous: each array has the same properties, and in particular, is associated to the same Pencil configuration.

For convenience, certain operations defined for PencilArray are also defined for PencilArrayCollection, and return the same value as for a single PencilArray. Some examples are pencil, range_local and get_comm.

Also note that functions from Base, such as size or ndims, are not overloaded for PencilArrayCollection, since they already have a definition for tuples and arrays (and redefining them would be type piracy...).

PencilFFTs.PencilArrays.MaybePencilArrayCollection — Constant

MaybePencilArrayCollection

UnionAll type representing either a PencilArray or a collection of PencilArrays.

See also PencilArrayCollection.

PencilFFTs.PencilArrays.ManyPencilArray — Type

ManyPencilArray{T,N,M}

Container holding M different PencilArray views to the same underlying data buffer. All views share the same element type T and dimensionality N.

This can be useful to perform in-place operations on PencilArray data.

ManyPencilArray(pencils...; extra_dims=())

Create a ManyPencilArray container that can hold data associated to all the given Pencils.

The optional extra_dims argument is the same as for PencilArray.

PencilArray methods

PencilFFTs.PencilArrays.extra_dims — Method

extra_dims(x::PencilArray)
extra_dims(x::PencilArrayCollection)

Return tuple with size of "extra" dimensions of PencilArray.

PencilFFTs.PencilArrays.MPITopologies.get_comm — Method

get_comm(x::PencilArray)
get_comm(x::PencilArrayCollection)

Get MPI communicator associated to a pencil-distributed array.

PencilFFTs.PencilArrays.get_permutation — Method

get_permutation(x::PencilArray)
get_permutation(x::PencilArrayCollection)

Get index permutation associated to the given PencilArray.

Returns NoPermutation() if there is no associated permutation.

PencilFFTs.PencilArrays.global_view — Method

global_view(x::PencilArray)

Create an OffsetArray of a PencilArray that takes unpermuted global indices.

PencilFFTs.PencilArrays.ndims_extra — Method

ndims_extra(x::PencilArray)
ndims_extra(x::PencilArrayCollection)

Number of "extra" dimensions associated to PencilArray.

These are the dimensions that are not associated to the domain geometry. For instance, they may correspond to vector or tensor components.

These dimensions correspond to the rightmost indices of the array.

The total number of dimensions of a PencilArray is given by:

ndims(x) == ndims_space(x) + ndims_extra(x)

PencilFFTs.PencilArrays.ndims_space — Method

ndims_space(x::PencilArray)
ndims_space(x::PencilArrayCollection)

Number of dimensions associated to the domain geometry.

These dimensions correspond to the leftmost indices of the array.

The total number of dimensions of a PencilArray is given by:

ndims(x) == ndims_space(x) + ndims_extra(x)

Base.parent — Method

parent(x::PencilArray)

Return array wrapped by a PencilArray.

PencilFFTs.PencilArrays.pencil — Method

pencil(x::PencilArray)

Return decomposition configuration associated to a PencilArray.

Base.pointer — Method

pointer(x::PencilArray)

Return pointer to the start of the underlying data.

Use with caution: this may not make a lot of sense if the underlying data is not contiguous or strided (e.g. if the PencilArray is wrapping a non-strided SubArray).

PencilFFTs.PencilArrays.range_local — Method

range_local(x::PencilArray; permute=false)
range_local(x::PencilArrayCollection; permute=false)

Local data range held by the PencilArray.

By default the dimensions are not permuted, i.e. they are returned in logical order.

Base.size — Method

size(x::PencilArray)

Return logical (unpermuted) local dimensions of a PencilArray.

PencilFFTs.PencilArrays.size_local — Method

size_local(x::PencilArray; permute=false)
size_local(x::PencilArrayCollection; permute=false)

Local dimensions of the data held by the PencilArray.

If permute=false, this is the same as size(x).

See also size_local(::Pencil).

PencilFFTs.PencilArrays.size_global — Method

size_global(x::PencilArray; permute=false)
size_global(x::PencilArrayCollection; permute=false)

Global dimensions associated to the given array.

By default, the logical (unpermuted) dimensions of the dataset are returned.

See also size_global(::Pencil).

ManyPencilArray methods

Base.first — Method

first(A::ManyPencilArray)

Returns the first PencilArray wrapped by A.

Base.getindex — Method

getindex(A::ManyPencilArray, ::Val{i})
getindex(A::ManyPencilArray, i::Integer)

Returns the i-th PencilArray wrapped by A.

If possible, the Val{i} form should be preferred, as it is more efficient and it allows the compiler to know the return type.

See also first(::ManyPencilArray), last(::ManyPencilArray).

Example

A = ManyPencilArray(pencil1, pencil2, pencil3)

# Get the PencilArray associated to `pencil2`.
# u2 = A[2]
u2 = A[Val(2)]  # faster!

Base.last — Method

last(A::ManyPencilArray)

Returns the last PencilArray wrapped by A.

Base.length — Method

length(A::ManyPencilArray)

Returns the number of PencilArrays wrapped by A.

Global MPI operations

One of the most time-consuming parts of a large-scale computation involving multidimensional FFTs, is the global data transpositions between different MPI decomposition configurations. In PencilArrays, this is performed by the transpose! function, which takes two PencilArrays, typically associated to two different configurations. The implementation performs comparably to similar implementations in lower-level languages (see Benchmarks).

Also provided is a gather function that creates a single global array from decomposed data. This can be useful for tests (in fact, it is used in the PencilArrays tests to verify the correctness of the transpositions), but shouldn't be used with large datasets. It is generally useful for small problems where the global size of the data can easily fit the locally available memory.

PencilFFTs.PencilArrays.Transpositions.Transposition — Type

Transposition

Holds data for transposition between two pencil configurations.

Transposition(dest::PencilArray{T,N}, src::PencilArray{T,N};
              method=Transpositions.IsendIrecv())

Prepare transposition of arrays from one pencil configuration to the other.

The two pencil configurations must be compatible for transposition:

they must share the same MPI Cartesian topology,

they must have the same global data size,

when written as a sorted tuple, the decomposed dimensions must be almost the same, with at most one difference. For instance, if the input of a 3D dataset is decomposed in (2, 3), then the output may be decomposed in (1, 3), but not in (1, 2). If the decomposed dimensions are the same, then no transposition is performed, and data is just copied if needed.

The src and dest arrays may be aliased (they can share memory space).

Performance tuning

The method argument allows to choose between transposition implementations. This can be useful to tune performance of MPI data transfers. Two values are currently accepted:

Transpositions.IsendIrecv() uses non-blocking point-to-point data transfers (MPI_Isend and MPI_Irecv). This may be more performant since data transfers are interleaved with local data transpositions (index permutation of received data). This is the default.

Transpositions.Alltoallv() uses MPI_Alltoallv for global data transpositions.

LinearAlgebra.transpose! — Function

transpose!(t::Transposition; waitall=true)
transpose!(dest::PencilArray{T,N}, src::PencilArray{T,N};
           method=Transpositions.IsendIrecv())

Transpose data from one pencil configuration to the other.

The first variant allows to optionally delay the wait for MPI send operations to complete. This is useful if the caller wants to perform other operations with the already received data. To do this, the caller should pass waitall=false, and manually invoke MPI.Waitall! on the Transposition object once the operations are done. Note that this option only has an effect when the transposition method is IsendIrecv.

See Transposition for details.

MPI.Waitall! — Function

MPI.Waitall!(t::Transposition)

Wait for completion of all unfinished MPI communications related to the transposition.

PencilFFTs.PencilArrays.gather — Function

gather(x::PencilArray, [root::Integer=0])

Gather data from all MPI processes into one (big) array.

Data is received by the root process.

Returns the full array on the root process, and nothing on the other processes.

This can be useful for testing, but it shouldn't be used with very large datasets!

Measuring performance

It is possible to measure the time spent in different sections of the MPI data transposition routines using the TimerOutputs package. This has a (very small) performance overhead, so it is disabled by default. To enable time measurements, call TimerOutputs.enable_debug_timings after loading PencilFFTs (see below for an example). For more details see the TimerOutputs docs.

Minimal example:

using MPI
using PencilFFTs.PencilArrays
using TimerOutputs

# Enable timing of `PencilArrays` functions
TimerOutputs.enable_debug_timings(PencilArrays)
TimerOutputs.enable_debug_timings(Transpositions)

MPI.Init()

pencil = Pencil(#= args... =#)

# [do stuff with `pencil`...]

# Retrieve and print timing data associated to `plan`
to = get_timer(pencil)
print_timer(to)

By default, each Pencil has its own TimerOutput. If you already have a TimerOutput, you can pass it to the Pencil constructor:

to = TimerOutput()
pencil = Pencil(..., timer=to)

# [do stuff with `pencil`...]

print_timer(to)

Index

PencilFFTs.PencilArrays

PencilFFTs.PencilArrays.MPITopologies.MPITopology

PencilFFTs.PencilArrays.ManyPencilArray

PencilFFTs.PencilArrays.NoPermutation

PencilFFTs.PencilArrays.Pencil

PencilFFTs.PencilArrays.PencilArray

PencilFFTs.PencilArrays.Permutation

PencilFFTs.PencilArrays.Transpositions.Transposition

PencilFFTs.PencilArrays.MaybePencilArrayCollection

PencilFFTs.PencilArrays.PencilArrayCollection

Base.eltype

Base.first

Base.getindex

Base.last

Base.length

Base.length

Base.length

Base.ndims

Base.ndims

Base.parent

Base.pointer

Base.size

Base.size

LinearAlgebra.transpose!

MPI.Waitall!

PencilFFTs.PencilArrays.MPITopologies.get_comm

PencilFFTs.PencilArrays.MPITopologies.get_comm

PencilFFTs.PencilArrays.MPITopologies.get_comm

PencilFFTs.PencilArrays.extra_dims

PencilFFTs.PencilArrays.gather

PencilFFTs.PencilArrays.get_decomposition

PencilFFTs.PencilArrays.get_permutation

PencilFFTs.PencilArrays.get_permutation

PencilFFTs.PencilArrays.global_view

PencilFFTs.PencilArrays.ndims_extra

PencilFFTs.PencilArrays.ndims_space

PencilFFTs.PencilArrays.pencil

PencilFFTs.PencilArrays.range_local

PencilFFTs.PencilArrays.range_local

PencilFFTs.PencilArrays.size_global

PencilFFTs.PencilArrays.size_global

PencilFFTs.PencilArrays.size_local

PencilFFTs.PencilArrays.size_local

PencilFFTs.PencilArrays.to_local

1Why would we want this? Perhaps because we want to efficiently perform FFTs along $y$, which, under this permutation, would be the fastest dimension.