PencilFFTs module

Types

PencilFFTs.PencilFFTPlan — Type

PencilFFTPlan{T,N,M}

Plan for N-dimensional FFT-based transform on MPI-distributed data.

PencilFFTPlan(size_global::Dims{N}, transforms,
              proc_dims::Dims{M}, comm::MPI.Comm, [real_type=Float64];
              fftw_flags=FFTW.ESTIMATE, fftw_timelimit=FFTW.NO_TIMELIMIT,
              permute_dims=Val(true),
              transpose_method=TransposeMethods.IsendIrecv(),
              timer=TimerOutput(),
              )

Create plan for N-dimensional transform.

size_global specifies the global dimensions of the input data.

transforms should be a tuple of length N specifying the transforms to be applied along each dimension. Each element must be a subtype of Transforms.AbstractTransform. For all the possible transforms, see Transform types. Alternatively, transforms may be a single transform that will be automatically expanded into N equivalent transforms. This is illustrated in the example below.

The transforms are applied one dimension at a time, with the leftmost dimension first for forward transforms. For multidimensional transforms of real data, this means that a real-to-complex transform must be performed along the first dimension, and then complex-to-complex transforms are performed along the other two dimensions (see example below).

The data is distributed over the MPI processes in the comm communicator. The distribution is performed over M dimensions (with M < N) according to the values in proc_dims, which specifies the number of MPI processes to put along each dimension.

Optional arguments

The floating point precision can be selected by setting real_type parameter, which is Float64 by default.

The keyword arguments fftw_flags and fftw_timelimit are passed to the FFTW plan creation functions (see AbstractFFTs docs).

permute_dims determines whether the indices of the output data should be reversed. For instance, if the input data has global dimensions (Nx, Ny, Nz), then the output of a complex-to-complex FFT would have dimensions (Nz, Ny, Nx). This enables FFTs to always be performed along the first (i.e. fastest) array dimension, which could lead to performance gains. This option is enabled by default. For type inference reasons, it must be a value type (Val(true) or Val(false)).

transpose_method allows to select between implementations of the global data transpositions. See the transpose! docs for details.

timer should be a TimerOutput object. See Measuring performance for details.

Example

Suppose we want to perform a 3D transform of real data. The data is to be decomposed along two dimensions, over 8 MPI processes:

size_global = (64, 32, 128)  # size of real input data

# Perform real-to-complex transform along the first dimension, then
# complex-to-complex transforms along the other dimensions.
transforms = (Transforms.RFFT(), Transforms.FFT(), Transforms.FFT())
# transforms = Transforms.RFFT()  # this is equivalent to the above line

proc_dims = (4, 2)  # 2D decomposition
comm = MPI.COMM_WORLD

plan = PencilFFTPlan(size_global, transforms, proc_dims, comm)

Functions

PencilFFTs.allocate_input — Function

allocate_input(p::PencilFFTPlan)

Allocate uninitialised distributed array that can hold input data for the given plan.

PencilFFTs.allocate_output — Function

allocate_output(p::PencilFFTPlan)

Allocate uninitialised distributed array that can hold output data for the given plan.

PencilFFTs.Pencils.get_timer — Function

get_timer(p::Pencil)

Get TimerOutput attached to a Pencil.

See Measuring performance for details.

get_timer(p::PencilFFTPlan)

Get TimerOutput attached to a PencilFFTPlan.

See Measuring performance for details.

Measuring performance

It is possible to measure the time spent in different sections of the distributed transforms using the TimerOutput package. This has a (very small) performance overhead, so it is disabled by default. To enable time measurements, call TimerOutputs.enable_debug_timings(PencilFFTs) and TimerOutputs.enable_debug_timings(PencilFFTs.Pencils) after loading PencilFFTs. For more details see the TimerOutput docs.

Minimal example:

using MPI
using PencilFFTs
using TimerOutput

# Enable timing of `PencilFFTs` functions
TimerOutputs.enable_debug_timings(PencilFFTs)
TimerOutputs.enable_debug_timings(PencilFFTs.Pencils)

MPI.Init()

plan = PencilFFTPlan(#= args... =#)

# [do stuff with `plan`...]

# Retrieve and print timing data associated to `plan`
to = get_timer(plan)
print_timer(to)

By default, each PencilFFTPlan has its own TimerOutput. If you already have a TimerOutput, you can pass it to the PencilFFTPlan constructor:

to = TimerOutput()
plan = PencilFFTPlan(..., timer=to)

# [do stuff with `plan`...]

print_timer(to)

Devdocs

PencilFFTs.GlobalFFTParams — Type

GlobalFFTParams{T, N}

Specifies the global parameters for an N-dimensional distributed transform. These include the element type T and global data sizes of input and output data, as well as the transform types to be performed along each dimension.

GlobalFFTParams(size_global, transforms, [real_type=Float64])

Define parameters for N-dimensional transform.

transforms must be a tuple of length N specifying the transforms to be applied along each dimension. Each element must be a subtype of Transforms.AbstractTransform. For all the possible transforms, see Transform types.

The element type must be a real type accepted by FFTW, i.e. either Float32 or Float64.

Note that the transforms are applied one dimension at a time, with the leftmost dimension first for forward transforms.

Example

To perform a 3D FFT of real data, first a real-to-complex FFT must be applied along the first dimension, followed by two complex-to-complex FFTs along the other dimensions:

julia> size_global = (64, 32, 128);  # size of real input data

julia> transforms = (Transforms.RFFT(), Transforms.FFT(), Transforms.FFT());

julia> fft_params = PencilFFTs.GlobalFFTParams(size_global, transforms);