Parallel I/O

The PencilFFTs.PencilIO module contains functions for saving and loading PencilArrays to disk using MPI-IO. Parallel I/O is performed using Parallel HDF5 through the HDF5.jl package.

The implemented approach consists in storing the data coming from different MPI processes in a single file. This strategy scales better in terms of number of files, and is more convenient, than that of storing one file per process. However, the performance is very sensitive to the configuration of the underlying file system. In distributed file systems such as Lustre, it is worth tuning parameters such as the stripe count and stripe size (see for instance the Parallel HDF5 page for more information).

Setting-up Parallel HDF5

Parallel HDF5 is not enabled in the default installation of HDF5.jl. For Parallel HDF5 to work, the HDF5 C libraries wrapped by HDF5.jl must be compiled with parallel support and linked to the specific MPI implementation that will be used for parallel I/O. HDF5.jl must be explicitly instructed to use parallel-enabled HDF5 libraries available in the system. Similarly, MPI.jl must be instructed to use the corresponding MPI libraries. This is detailed in the sections below.

Parallel-enabled HDF5 libraries are usually included in computing clusters and linked to the available MPI implementations. They are also available via the package manager of a number of Linux distributions. (For instance, Fedora includes the hdf5-mpich-devel and hdf5-openmpi-devel packages, respectively linked to the MPICH and OpenMPI libraries in the Fedora repositories.)

The following step-by-step guide assumes one already has access to parallel-enabled HDF5 libraries linked to an existent MPI installation.

1. Using system-provided MPI libraries

Set the environment variable JULIA_MPI_BINARY=system and then run ]build MPI from Julia. For more control, one can also set the JULIA_MPI_PATH environment variable to the top-level installation directory of the MPI library.

See the MPI.jl docs for details.

2. Using parallel HDF5 libraries

Set the JULIA_HDF5_LIBRARY_PATH environment variable to the directory where the HDF5 libraries compiled with parallel support are found. Then run ]build HDF5 from Julia. Note that the selected HDF5 library must be linked to the MPI library chosen in the previous section. For the set-up to be persistent across HDF5.jl updates, consider setting JULIA_HDF5_LIBRARY_PATH in ~/.bashrc or similar.

See the HDF5.jl README for details.

3. Loading PencilIO

In the PencilIO module, the HDF5.jl package is lazy-loaded using Requires. This means that, in Julia code, PencilFFTs must be loaded after HDF5 for parallel I/O functionality to be available.

The following order of usings ensures that parallel I/O support is available:

using MPI
using HDF5
using PencilFFTs

Library

PencilFFTs.PencilIO.ph5open — Function

ph5open([f::Function], filename, [mode="r"], comm::MPI.Comm,
        [info::MPI.Info=MPI.Info()], prop_lists...) -> HDF5.File

Open parallel HDF5 file.

This function is a thin wrapper over HDF5.h5open. It converts MPI.jl types (MPI.Comm and MPI.Info) to their counterparts in HDF5.jl. It also throws an informative error if the loaded HDF5 libraries do not include parallel support.

Property lists

This function automatically sets the fapl_mpio file access property list to the given MPI communicator and info object. Other property lists should be given as name-value pairs, following the h5open syntax.

Property lists are passed to h5f_create. The following property types are recognised:

file creation properties,

file access properties.

Base.setindex! — Function

setindex!(g::Union{HDF5File,HDF5Group}, x::MaybePencilArrayCollection,
          name::String, prop_lists...; chunks=false, collective=true)

Write PencilArray or PencilArrayCollection to parallel HDF5 file.

For performance reasons, the memory layout of the data is conserved. In other words, if the dimensions of a PencilArray are permuted in memory, then the data is written in permuted form.

In the case of a PencilArrayCollection, each array of the collection is written as a single component of a higher-dimension dataset.

Optional arguments

if chunks=true, data is written in chunks, with roughly one chunk per MPI process. This may (or may not) improve performance in parallel filesystems.

if collective=true, the dataset is written collectivelly. This is usually recommended for performance.

additional property lists may be specified by name-value pairs in prop_lists, following the HDF5.jl syntax. These property lists take precedence over keyword arguments. For instance, if the "dxpl_mpio", HDF5.H5FD_MPIO_COLLECTIVE pair is passed, then the value of the collective argument is ignored.

Property lists

Property lists are passed to h5d_create and h5d_write. The following property types are recognised:

link creation properties,

dataset creation properties,

dataset access properties,

dataset transfer properties.

Example

Open a parallel HDF5 file and write some PencilArrays to the file:

u = PencilArray(...)
v = PencilArray(...)

comm = get_comm(u)
info = MPI.Info()

ph5open("filename.h5", "w", comm, info) do ff
    ff["u", chunks=true] = u
    ff["uv"] = (u, v)  # this is a two-component PencilArrayCollection (assuming equal dimensions of `u` and `v`)
end

Base.read! — Function

read!(g::Union{HDF5File,HDF5Group}, x::MaybePencilArrayCollection,
      name::String, prop_lists...; collective=true)

Read PencilArray or PencilArrayCollection from parallel HDF5 file.

See setindex! for details on optional arguments.

Property lists

Property lists are passed to h5d_open and h5d_read. The following property types are recognised:

dataset access properties,

dataset transfer properties.

Example

Open a parallel HDF5 file and read some PencilArrays:

u = PencilArray(...)
v = PencilArray(...)

comm = get_comm(u)
info = MPI.Info()

ph5open("filename.h5", "r", comm, info) do ff
    read!(ff, u, "u")
    read!(ff, (u, v), "uv")
end

PencilFFTs.PencilIO.hdf5_has_parallel — Function

hdf5_has_parallel() -> Bool

Returns true if the loaded HDF5 libraries support MPI I/O.