Pencils module

Pencils module

PencilFFTs.Pencils — Module.

Module for multidimensional data decomposition using MPI.

Handles different decomposition configurations and data transpositions between them. Also defines relevant data structures for handling distributed data.

source

Types

PencilFFTs.Pencils.MPITopology — Type.

MPITopology{N}

Describes an N-dimensional Cartesian MPI decomposition topology.

MPITopology(comm::MPI.Comm, pdims::Dims{N}) where N

Create N-dimensional MPI topology information.

The pdims tuple specifies the number of MPI processes to put in every dimension of the topology. The product of its values must be equal to the number of processes in communicator comm.

Example

# Divide 2D topology into 4×2 blocks.
comm = MPI.COMM_WORLD
@assert MPI.Comm_size(comm) == 8
topology = MPITopology(comm, (4, 2))

MPITopology{N}(comm_cart::MPI.Comm) where N

Create topology information from MPI communicator with Cartesian topology. The topology must have dimension N.

source

PencilFFTs.Pencils.Pencil — Type.

Pencil{N,M,T}

Describes the decomposition of an N-dimensional Cartesian geometry among MPI processes along M directions (with M < N).

The Pencil describes the decomposition of arrays of element type T.

Pencil(topology::MPITopology{M}, size_global::Dims{N},
       decomp_dims::Dims{M}, [element_type=Float64];
       permute=nothing, timer=TimerOutput())

Define the decomposition of an N-dimensional geometry along M dimensions.

The dimensions of the geometry are given by size_global = (N1, N2, ...). The Pencil describes the decomposition of an array of dimensions size_global and type T across a group of MPI processes.

Data is distributed over the given M-dimensional MPI topology (with M < N). The decomposed dimensions are given by decomp_dims.

The optional parameter perm should be a tuple defining a permutation of the data indices. This may be useful for performance reasons, since it may be preferable (e.g. for FFTs) that the data is contiguous along the pencil direction.

It is also possible to pass a TimerOutput to the constructor. See Measuring performance for details.

Examples

Decompose a 3D geometry of global dimensions $N_x × N_y × N_z = 4×8×12$ along the second ($y$) and third ($z$) dimensions.

Pencil(topology, (4, 8, 12), (2, 3))             # data is in (x, y, z) order
Pencil(topology, (4, 8, 12), (2, 3), (3, 2, 1))  # data is in (z, y, x) order

In the second case, the actual data is stored in (z, y, x) order within each MPI process.

Pencil(p::Pencil{N,M}, [element_type=eltype(p)];
       decomp_dims::Dims{M}=get_decomposition(p),
       size_global::Dims{N}=size_global(p),
       permute::P=get_permutation(p),
       timer::TimerOutput=get_timer(p))

Create new pencil configuration from an existent one.

This constructor enables sharing temporary data buffers between the two pencil configurations, leading to reduced global memory usage.

source

PencilFFTs.Pencils.PencilArray — Type.

PencilArray(pencil::P, data::AbstractArray{T,N})

Create array wrapper with pencil decomposition information.

The array dimensions and element type must be consistent with those of the given pencil.

The data array can have one or more extra dimensions to the left (fast indices). For instance, these may correspond to vector or tensor components.

Example

Suppose pencil has local dimensions (10, 20, 30). Then:

PencilArray(pencil, zeros(10, 20, 30))        # works (scalar)
PencilArray(pencil, zeros(3, 10, 20, 30))     # works (3-component vector)
PencilArray(pencil, zeros(4, 3, 10, 20, 30))  # works (4×3 tensor)
PencilArray(pencil, zeros(10, 20, 30, 3))     # fails

PencilArray(pencil::Pencil, [extra_dims...])

Allocate uninitialised PencilArray that can hold data in the local pencil.

Extra dimensions, for instance representing vector components, can be specified. These dimensions are added to the leftmost (fastest) indices of the resulting array.

Example

Suppose pencil has local dimensions (10, 20, 30). Then:

PencilArray(pencil)        # array dimensions are (10, 20, 30)
PencilArray(pencil, 4, 3)  # array dimensions are (4, 3, 10, 20, 30)

source

Functions

PencilFFTs.Pencils.data — Function.

data(x::PencilArray)

Returns array wrapped by the PencilArray.

source

Base.eltype — Function.

eltype(p::Pencil)

Element type associated to the given pencil.

source

PencilFFTs.Pencils.gather — Function.

gather(x::PencilArray, [root::Integer=0])

Gather data from all MPI processes into one (big) array.

Data is received by the root process.

Returns the full array on the root process, and nothing on the other processes.

This can be useful for testing, but it shouldn't be used with very large datasets!

source

PencilFFTs.Pencils.get_comm — Function.

get_comm(t::MPITopology)

Get MPI communicator associated to an MPI Cartesian topology.

source

get_comm(p::Pencil)

Get MPI communicator associated to an MPI decomposition scheme.

source

get_comm(x::PencilArray)

Get MPI communicator associated to a pencil-distributed array.

source

get_comm(p::PencilFFTPlan)

Get MPI communicator associated to a PencilFFTPlan.

source

PencilFFTs.Pencils.get_decomposition — Function.

get_decomposition(p::Pencil)

Get tuple with decomposed dimensions of the given pencil configuration.

source

PencilFFTs.Pencils.get_permutation — Function.

get_permutation(p::Pencil)

Get index permutation associated to the given pencil configuration.

Returns nothing if there is no associated permutation.

source

Base.length — Function.

length(t::MPITopology)

Get total size of Cartesian topology (i.e. total number of MPI processes).

source

Base.ndims — Function.

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

source

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

source

PencilFFTs.Pencils.ndims_extra — Function.

ndims_extra(x::PencilArray)

Number of "extra" dimensions associated to PencilArray.

These are the dimensions that are not associated to the domain geometry. For instance, they may correspond to vector or tensor components.

source

Base.parent — Function.

parent(x::PencilArray)

Returns the actual array containing the PencilArray data.

If the PencilArray is wrapping a SubArray, then this returns its "parent array".

source

PencilFFTs.Pencils.pencil — Function.

pencil(x::PencilArray)

Returns decomposition configuration associated to the PencilArray.

source

Base.size — Function.

size(t::MPITopology)

Get dimensions of Cartesian topology.

source

PencilFFTs.Pencils.size_global — Function.

size_global(p::Pencil)

Global dimensions of the Cartesian grid associated to the given domain decomposition.

Unlike size_local, the returned dimensions are not permuted to match the dimensions of the local data.

source

size_global(x::PencilArray)

Global dimensions associated to the given array.

Unlike size, the returned dimensions are not permuted according to the associated pencil configuration.

source

PencilFFTs.Pencils.size_local — Function.

size_local(p::Pencil; permute=true)

Local dimensions of the Cartesian grid held by the pencil.

By default the dimensions are permuted to match those of the associated data arrays.

source

PencilFFTs.Pencils.to_local — Function.

to_local(p::Pencil, global_inds; permute=true)

Convert non-permuted global indices to local indices.

Indices are permuted by default using the permutation associated to the pencil configuration p.

source

LinearAlgebra.transpose! — Function.

transpose!(dest::PencilArray{T,N}, src::PencilArray{T,N};
           method=TransposeMethods.IsendIrecv())

Transpose data from one pencil configuration to the other.

The two pencil configurations must be compatible for transposition:

they must share the same MPI Cartesian topology,

they must have the same global data size,

when written as a sorted tuple, the decomposed dimensions must be almost the same, with at most one difference. For instance, if the input of a 3D dataset is decomposed in (2, 3), then the output may be decomposed in (1, 3), but not in (1, 2). If the decomposed dimensions are the same, then no transposition is performed, and data is just copied if needed.

Performance tuning

The method argument allows to choose between transposition implementations. This can be useful to tune performance of MPI data transfers. Two values are currently accepted:

TransposeMethods.IsendIrecv() uses non-blocking point-to-point data transfers (MPI_Isend and MPI_Irecv). This may be more performant since data transfers are interleaved with local data transpositions (index permutation of received data). This is the default.

TransposeMethods.Alltoallv() uses MPI_Alltoallv for global data transpositions.

source

Measuring performance

It is possible to measure the time spent in different sections of the MPI data transposition routines using the TimerOutput package. This has a (very small) performance overhead, so it is disabled by default. To enable time measurements, call TimerOutputs.enable_debug_timings(PencilFFTs.Pencils) after loading PencilFFTs. For more details see the TimerOutput docs.

Minimal example:

using MPI
using PencilFFTs.Pencils
using TimerOutput

# Enable timing of `Pencils` functions
TimerOutputs.enable_debug_timings(PencilFFTs.Pencils)

MPI.Init()

pencil = Pencil(#= args... =#)

# [do stuff with `pencil`...]

# Retrieve and print timing data associated to `plan`
to = get_timer(pencil)
print_timer(to)

By default, each Pencil has its own TimerOutput. If you already have a TimerOutput, you can pass it to the Pencil constructor:

to = TimerOutput()
pencil = Pencil(..., timer=to)

# [do stuff with `pencil`...]

print_timer(to)