MPI-distributed data

The distribution of global data among MPI processes is managed by the Pencils module. This module may be used independently of the FFT functionality.

The Pencils module defines types that describe an MPI Cartesian topology and the decomposition of data over MPI processes. Also, array wrappers allow to conveniently (and efficiently) deal with the MPI-decomposed data.

PencilFFTs.Pencils — Module

Module for multidimensional data decomposition using MPI.

Handles different decomposition configurations and data transpositions between them. Also defines relevant data structures for handling distributed data.

MPI topology

The MPITopology type defines the MPI Cartesian topology of the decomposition. In other words, it contains information about the number of decomposed dimensions, and the number of processes in each of these dimensions.

At the lower level, MPITopology uses MPI_Cart_create to define a Cartesian MPI communicator.

PencilFFTs.Pencils.MPITopologies.MPITopology — Type

MPITopology{N}

Describes an N-dimensional Cartesian MPI decomposition topology.

MPITopology(comm::MPI.Comm, pdims::Dims{N}) where N

Create N-dimensional MPI topology information.

The pdims tuple specifies the number of MPI processes to put in every dimension of the topology. The product of its values must be equal to the number of processes in communicator comm.

Example

# Divide 2D topology into 4×2 blocks.
comm = MPI.COMM_WORLD
@assert MPI.Comm_size(comm) == 8
topology = MPITopology(comm, (4, 2))

MPITopology{N}(comm_cart::MPI.Comm) where N

Create topology information from MPI communicator with Cartesian topology (typically constructed using MPI.Cart_create). The topology must have dimension N.

PencilFFTs.Pencils.MPITopologies.get_comm — Method

get_comm(t::MPITopology)

Get MPI communicator associated to an MPI Cartesian topology.

Base.length — Method

length(t::MPITopology)

Get total size of Cartesian topology (i.e. total number of MPI processes).

Base.ndims — Method

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

Base.size — Method

size(t::MPITopology)

Get dimensions of Cartesian topology.

Pencil configurations

The Pencil type describes a given MPI decomposition configuration of multidimensional data.

PencilFFTs.Pencils.Pencil — Type

Pencil{N,M,T}

Describes the decomposition of an N-dimensional Cartesian geometry among MPI processes along M directions (with M < N).

The Pencil describes the decomposition of arrays of element type T.

Pencil(topology::MPITopology{M}, size_global::Dims{N},
       decomp_dims::Dims{M}, [element_type=Float64];
       permute::Union{Nothing, Val}=nothing,
       timer=TimerOutput())

Define the decomposition of an N-dimensional geometry along M dimensions.

The dimensions of the geometry are given by size_global = (N1, N2, ...). The Pencil describes the decomposition of an array of dimensions size_global and type T across a group of MPI processes.

Data is distributed over the given M-dimensional MPI topology (with M < N). The decomposed dimensions are given by decomp_dims.

The optional parameter perm should be a (compile-time) tuple defining a permutation of the data indices. Such permutation may be useful for performance reasons, since it may be preferable (e.g. for FFTs) that the data is contiguous along the pencil direction.

It is also possible to pass a TimerOutput to the constructor. See Measuring performance for details.

Examples

Decompose a 3D geometry of global dimensions $N_x × N_y × N_z = 4×8×12$ along the second ($y$) and third ($z$) dimensions.

Pencil(topology, (4, 8, 12), (2, 3))                          # data is in (x, y, z) order
Pencil(topology, (4, 8, 12), (2, 3), permute=Val((3, 2, 1)))  # data is in (z, y, x) order

In the second case, the actual data is stored in (z, y, x) order within each MPI process.

Pencil(p::Pencil{N,M}, [element_type=eltype(p)];
       decomp_dims::Dims{M}=get_decomposition(p),
       size_global::Dims{N}=size_global(p),
       permute::P=get_permutation(p),
       timer::TimerOutput=get_timer(p))

Create new pencil configuration from an existent one.

This constructor enables sharing temporary data buffers between the two pencil configurations, leading to reduced global memory usage.

Functions

Base.eltype — Method

eltype(Pencil)
eltype(p::Pencil)

Element type associated to the given pencil type.

PencilFFTs.Pencils.MPITopologies.get_comm — Method

get_comm(p::Pencil)

Get MPI communicator associated to an MPI decomposition scheme.

PencilFFTs.Pencils.get_decomposition — Method

get_decomposition(p::Pencil)

Get tuple with decomposed dimensions of the given pencil configuration.

PencilFFTs.Pencils.get_permutation — Method

get_permutation(p::Pencil)

Get index permutation associated to the given pencil configuration.

Returns nothing if there is no associated permutation.

Base.ndims — Method

ndims(t::MPITopology)

Get dimensionality of Cartesian topology.

ndims(p::Pencil)

Number of spatial dimensions associated to pencil data.

This corresponds to the total number of dimensions of the space, which includes the decomposed and non-decomposed dimensions.

PencilFFTs.Pencils.range_local — Method

range_local(p::Pencil; permute=true)

Local data range held by the pencil.

By default the dimensions are permuted to match those of the associated data arrays.

PencilFFTs.Pencils.size_global — Method

size_global(p::Pencil; permute=false)

Global dimensions of the Cartesian grid associated to the given domain decomposition.

Unlike size_local, by default the returned dimensions are not permuted to match the dimensions of the local data.

PencilFFTs.Pencils.size_local — Method

size_local(p::Pencil; permute=true)

Local dimensions of the data held by the pencil.

By default the dimensions are permuted to match those of the associated data arrays.

Array wrappers

Types

PencilFFTs.Pencils.PencilArray — Type

PencilArray(pencil::Pencil, data::AbstractArray{T,N})

Create array wrapper with pencil decomposition information.

The array dimensions and element type must be consistent with those of the given pencil.

Index permutations

If the Pencil has an associated index permutation, then data must have its dimensions permuted accordingly.

Unlike data, the resulting PencilArray should be accessed with unpermuted indices.

Example

Suppose pencil has local dimensions (10, 20, 30) before permutation, and has an asociated permutation (2, 3, 1). Then:

data = zeros(20, 30, 10)       # parent array (with permuted dimensions)

u = PencilArray(pencil, data)  # wrapper with dimensions (10, 20, 30)
@assert size(u) === (10, 20, 30)

u[15, 25, 5]                   # BoundsError (15 > 10 and 25 > 20)
u[5, 15, 25]                   # correct
parent(u)[15, 25, 5]           # correct

Extra dimensions

The data array can have one or more extra dimensions to the right (slow indices). For instance, these may correspond to vector or tensor components. These dimensions are not affected by permutations.

Example

dims = (20, 30, 10)
PencilArray(pencil, zeros(dims...))        # works (scalar)
PencilArray(pencil, zeros(dims..., 3))     # works (3-component vector)
PencilArray(pencil, zeros(dims..., 4, 3))  # works (4×3 tensor)
PencilArray(pencil, zeros(3, dims...))     # fails

PencilArray(pencil::Pencil, [extra_dims=()])

Allocate uninitialised PencilArray that can hold data in the local pencil.

Extra dimensions, for instance representing vector components, can be specified. These dimensions are added to the rightmost (slowest) indices of the resulting array.

Example

Suppose pencil has local dimensions (20, 10, 30) after permutation. Then:

PencilArray(pencil)          # array dimensions are (20, 10, 30)
PencilArray(pencil, (4, 3))  # array dimensions are (20, 10, 30, 4, 3)

PencilFFTs.Pencils.PencilArrayCollection — Constant

PencilArrayCollection

UnionAll type describing a collection of PencilArrays.

Such a collection can be a tuple or an array of PencilArrays.

Collections are by assumption homogeneous: each array has the same properties, and in particular, is associated to the same Pencil configuration.

For convenience, certain operations defined for PencilArray are also defined for PencilArrayCollection, and return the same value as for a single PencilArray. Some examples are pencil, range_local and get_comm.

Also note that functions from Base, such as size or ndims, are not overloaded for PencilArrayCollection, since they already have a definition for tuples and arrays (and redefining them would be type piracy...).

PencilFFTs.Pencils.MaybePencilArrayCollection — Constant

MaybePencilArrayCollection

UnionAll type representing either a PencilArray or a collection of PencilArrays.

See also PencilArrayCollection.

Functions

PencilFFTs.Pencils.extra_dims — Method

extra_dims(x::PencilArray)
extra_dims(x::PencilArrayCollection)

Return tuple with size of "extra" dimensions of PencilArray.

PencilFFTs.Pencils.MPITopologies.get_comm — Method

get_comm(x::PencilArray)
get_comm(x::PencilArrayCollection)

Get MPI communicator associated to a pencil-distributed array.

PencilFFTs.Pencils.get_permutation — Method

get_permutation(x::PencilArray)
get_permutation(x::PencilArrayCollection)

Get index permutation associated to the given PencilArray.

Returns nothing if there is no associated permutation.

PencilFFTs.Pencils.global_view — Method

global_view(x::PencilArray)

Create an OffsetArray of a PencilArray that takes unpermuted global indices.

PencilFFTs.Pencils.ndims_extra — Method

ndims_extra(x::PencilArray)
ndims_extra(x::PencilArrayCollection)

Number of "extra" dimensions associated to PencilArray.

These are the dimensions that are not associated to the domain geometry. For instance, they may correspond to vector or tensor components.

These dimensions correspond to the rightmost indices of the array.

The total number of dimensions of a PencilArray is given by:

ndims(x) == ndims_space(x) + ndims_extra(x)

PencilFFTs.Pencils.ndims_space — Method

ndims_space(x::PencilArray)
ndims_space(x::PencilArrayCollection)

Number of dimensions associated to the domain geometry.

These dimensions correspond to the leftmost indices of the array.

The total number of dimensions of a PencilArray is given by:

ndims(x) == ndims_space(x) + ndims_extra(x)

Base.parent — Method

parent(x::PencilArray)

Return array wrapped by a PencilArray.

PencilFFTs.Pencils.pencil — Method

pencil(x::PencilArray)

Return decomposition configuration associated to a PencilArray.

PencilFFTs.Pencils.range_local — Method

range_local(x::PencilArray; permute=true)
range_local(x::PencilArrayCollection; permute=true)

Local data range held by the PencilArray.

By default the dimensions not permuted, matching the order of indices in the array.

Base.size — Method

size(x::PencilArray)

Return (unpermuted) local dimensions of a PencilArray.

PencilFFTs.Pencils.size_global — Method

size_global(x::PencilArray; permute=false)
size_global(x::PencilArrayCollection; permute=false)

Global dimensions associated to the given array.

Unlike size, by default the returned dimensions are not permuted according to the associated pencil configuration.

PencilFFTs.Pencils.spatial_indices — Method

spatial_indices(x::PencilArray)
spatial_indices(x::PencilArrayCollection)
spatial_indices(x::GlobalPencilArray)

Create a CartesianIndices to iterate over the local "spatial" dimensions of a pencil-decomposed array.

The "spatial" dimensions are those that may be decomposed (as opposed to the "extra" dimensions, which are not considered by this function).

Global MPI operations

One of the most time-consuming parts of a large-scale computation involving multidimensional FFTs, is the global data transpositions between different MPI decomposition configurations. In Pencils, this is performed by the transpose! function, which takes two PencilArrays, typically associated to two different configurations. The implementation performs comparably to similar implementations in lower-level languages (see Benchmarks).

Also provided is a gather function that creates a single global array from decomposed data. This can be useful for tests (in fact, it is used in the Pencils tests to verify the correctness of the transpositions), but shouldn't be used with large datasets. It is generally useful for small problems where the global size of the data can easily fit the locally available memory.

LinearAlgebra.transpose! — Function

transpose!(dest::PencilArray{T,N}, src::PencilArray{T,N};
           method=TransposeMethods.IsendIrecv())

Transpose data from one pencil configuration to the other.

The two pencil configurations must be compatible for transposition:

they must share the same MPI Cartesian topology,

they must have the same global data size,

when written as a sorted tuple, the decomposed dimensions must be almost the same, with at most one difference. For instance, if the input of a 3D dataset is decomposed in (2, 3), then the output may be decomposed in (1, 3), but not in (1, 2). If the decomposed dimensions are the same, then no transposition is performed, and data is just copied if needed.

Performance tuning

The method argument allows to choose between transposition implementations. This can be useful to tune performance of MPI data transfers. Two values are currently accepted:

TransposeMethods.IsendIrecv() uses non-blocking point-to-point data transfers (MPI_Isend and MPI_Irecv). This may be more performant since data transfers are interleaved with local data transpositions (index permutation of received data). This is the default.

TransposeMethods.Alltoallv() uses MPI_Alltoallv for global data transpositions.

PencilFFTs.Pencils.gather — Function

gather(x::PencilArray, [root::Integer=0])

Gather data from all MPI processes into one (big) array.

Data is received by the root process.

Returns the full array on the root process, and nothing on the other processes.

This can be useful for testing, but it shouldn't be used with very large datasets!

Measuring performance

It is possible to measure the time spent in different sections of the MPI data transposition routines using the TimerOutputs package. This has a (very small) performance overhead, so it is disabled by default. To enable time measurements, call TimerOutputs.enable_debug_timings(PencilFFTs.Pencils) after loading PencilFFTs. For more details see the TimerOutputs docs.

Minimal example:

using MPI
using PencilFFTs.Pencils
using TimerOutput

# Enable timing of `Pencils` functions
TimerOutputs.enable_debug_timings(PencilFFTs.Pencils)

MPI.Init()

pencil = Pencil(#= args... =#)

# [do stuff with `pencil`...]

# Retrieve and print timing data associated to `plan`
to = get_timer(pencil)
print_timer(to)

By default, each Pencil has its own TimerOutput. If you already have a TimerOutput, you can pass it to the Pencil constructor:

to = TimerOutput()
pencil = Pencil(..., timer=to)

# [do stuff with `pencil`...]

print_timer(to)

Internal documentation

PencilFFTs.Pencils.to_local — Method

to_local(p::Pencil, global_inds; permute=true)

Convert non-permuted global indices to local indices.

Indices are permuted by default using the permutation associated to the pencil configuration p.

Index

Modules

PencilFFTs.Pencils

Types

PencilFFTs.Pencils.MPITopologies.MPITopology

PencilFFTs.Pencils.Pencil

PencilFFTs.Pencils.PencilArray

Functions

Base.eltype

Base.length

Base.ndims

Base.ndims

Base.parent

Base.size

Base.size

LinearAlgebra.transpose!

PencilFFTs.Pencils.MPITopologies.get_comm

PencilFFTs.Pencils.MPITopologies.get_comm

PencilFFTs.Pencils.MPITopologies.get_comm

PencilFFTs.Pencils.extra_dims

PencilFFTs.Pencils.gather

PencilFFTs.Pencils.get_decomposition

PencilFFTs.Pencils.get_permutation

PencilFFTs.Pencils.get_permutation

PencilFFTs.Pencils.global_view

PencilFFTs.Pencils.ndims_extra

PencilFFTs.Pencils.ndims_space

PencilFFTs.Pencils.pencil

PencilFFTs.Pencils.range_local

PencilFFTs.Pencils.range_local

PencilFFTs.Pencils.size_global

PencilFFTs.Pencils.size_global

PencilFFTs.Pencils.size_local

PencilFFTs.Pencils.spatial_indices

PencilFFTs.Pencils.to_local