Available transforms

PencilFFTs.Transforms — Module

Defines different one-dimensional FFT-based transforms.

The transforms are all subtypes of an AbstractTransform type.

When possible, the names of the transforms are kept consistent with the functions exported by AbstractFFTs.jl and FFTW.jl.

Transform types

PencilFFTs.Transforms.FFT — Type

FFT()

Complex-to-complex FFT.

See also AbstractFFTs.fft.

PencilFFTs.Transforms.FFT! — Type

FFT!()

In-place version of FFT.

See also AbstractFFTs.fft!.

PencilFFTs.Transforms.BFFT — Type

BFFT()

Unnormalised backward complex-to-complex FFT.

Like AbstractFFTs.bfft, this transform is not normalised. To obtain the inverse transform, divide the output by the length of the transformed dimension.

See also AbstractFFTs.bfft.

PencilFFTs.Transforms.BFFT! — Type

BFFT()

In-place version of BFFT.

See also AbstractFFTs.bfft!.

PencilFFTs.Transforms.RFFT — Type

RFFT()

Real-to-complex FFT.

See also AbstractFFTs.rfft.

PencilFFTs.Transforms.BRFFT — Type

BRFFT()

Unnormalised inverse of RFFT.

To obtain the inverse transform, divide the output by the length of the transformed dimension (of the real output array).

See also AbstractFFTs.brfft.

PencilFFTs.Transforms.R2R — Type

R2R{kind}()

Real-to-real transform of type kind.

The possible values of kind are those described in the FFTW.r2r docs and the FFTW manual:

discrete cosine transforms: FFTW.REDFT00, FFTW.REDFT01, FFTW.REDFFT10, FFTW.REDFFT11

discrete sine transforms: FFTW.RODFT00, FFTW.RODFT01, FFTW.RODFFT10, FFTW.RODFFT11

discrete Hartley transform: FFTW.DHT

Note: half-complex format DFTs (FFTW.R2HC, FFTW.HC2R) are not currently supported.

PencilFFTs.Transforms.R2R! — Type

R2R!{kind}()

In-place version of R2R.

See also FFTW.r2r!.

PencilFFTs.Transforms.NoTransform — Type

NoTransform()

Identity transform.

Specifies that no transformation should be applied.

PencilFFTs.Transforms.NoTransform! — Type

NoTransform!()

In-place version of NoTransform.

Internals

What follows is used internally by PencilFFTs.

Types

PencilFFTs.Transforms.AbstractCustomPlan — Type

AbstractCustomPlan

Abstract type defining a custom plan, to be used as an alternative to FFTW plans (FFTW.FFTWPlan).

The only custom plan defined in this module is IdentityPlan. The user can define other custom plans that are also subtypes of AbstractCustomPlan.

Note that plan returns a subtype of either FFTW.FFTWPlan or AbstractCustomPlan.

PencilFFTs.Transforms.AbstractTransform — Type

AbstractTransform

Specifies a one-dimensional FFT-based transform.

PencilFFTs.Transforms.IdentityPlan — Type

IdentityPlan

Type of plan associated to NoTransform.

PencilFFTs.Transforms.IdentityPlan! — Type

IdentityPlan!

Type of plan associated to NoTransform!.

PencilFFTs.Transforms.Plan — Constant

Plan = Union{FFTW.FFTWPlan, AbstractCustomPlan}

Union type representing any plan returned by plan.

See also AbstractCustomPlan.

Functions

PencilFFTs.Transforms.plan — Function

plan(transform::AbstractTransform, A, [dims];
     flags=FFTW.ESTIMATE, timelimit=Inf)

Create plan to transform array A along dimensions dims.

If dims is not specified, all dimensions of A are transformed.

For FFT plans, this function wraps the AbstractFFTs.jl and FFTW.jl plan creation functions. For more details on the function arguments, see AbstractFFTs.plan_fft.

In particular, note that for BRFFT plans, this function also requires the length d of the transform output along the first transformed dimension. This is described in the AbstractFFTs.irfft docs.

PencilFFTs.Transforms.binv — Function

binv(transform::AbstractTransform)

Returns the backwards transform associated to the given transform.

The backwards transform returned by this function is not normalised. The normalisation factor for a given array can be obtained by calling scale_factor.

Example

julia> binv(Transforms.FFT())
BFFT

julia> binv(Transforms.BRFFT())
RFFT

PencilFFTs.Transforms.scale_factor — Function

scale_factor(transform::AbstractTransform, A, [dims])

Get factor required to normalise the given array after a transformation along dimensions dims (all dimensions by default).

The array A must have the dimensions of the transform output.

Important: the dimensions dims must be the same that were passed to plan.

Examples

julia> C = zeros(ComplexF32, 3, 4, 5);

julia> scale_factor(Transforms.FFT(), C)
60

julia> scale_factor(Transforms.BFFT(), C)
60

julia> scale_factor(Transforms.BFFT(), C, 2:3)
20

julia> R = zeros(Float64, 3, 4, 5);

julia> scale_factor(Transforms.BRFFT(), R, 2)
4

julia> scale_factor(Transforms.BRFFT(), R, 2:3)
20

This will fail because the output of RFFT is complex, and R is a real array:

julia> scale_factor(Transforms.RFFT(), R, 2:3)
ERROR: MethodError: no method matching scale_factor(::PencilFFTs.Transforms.RFFT, ::Array{Float64,3}, ::UnitRange{Int64})

PencilFFTs.Transforms.eltype_input — Function

eltype_input(transform::AbstractTransform, real_type<:AbstractFloat)

Determine input data type for a given transform given the floating point precision of the input data.

Some transforms, such as R2R and NoTransform, can take both real and complex data. For those kinds of transforms, Nothing is returned.

Example

julia> eltype_input(Transforms.FFT(), Float32)
Complex{Float32}

julia> eltype_input(Transforms.RFFT(), Float64)
Float64

julia> eltype_input(Transforms.R2R{FFTW.REDFT01}(), Float64)
Nothing

julia> eltype_input(Transforms.NoTransform(), Float64)
Nothing

PencilFFTs.Transforms.eltype_output — Function

eltype_output(transform::AbstractTransform, eltype_input)

Returns the output data type for a given transform given the input type.

Throws ArgumentError if the input data type is incompatible with the transform type.

Example

julia> eltype_output(Transforms.NoTransform(), Float32)
Float32

julia> eltype_output(Transforms.RFFT(), Float64)
Complex{Float64}

julia> eltype_output(Transforms.BRFFT(), ComplexF32)
Float32

julia> eltype_output(Transforms.FFT(), Float64)
ERROR: ArgumentError: invalid input data type for PencilFFTs.Transforms.FFT: Float64

PencilFFTs.Transforms.expand_dims — Function

expand_dims(transform::AbstractTransform, Val(N))

Expand a single multidimensional transform into one transform per dimension.

Example

# Expand a real-to-complex transform in 3 dimensions.
julia> expand_dims(Transforms.RFFT(), Val(3))
(RFFT, FFT, FFT)

julia> expand_dims(Transforms.BRFFT(), Val(3))
(BRFFT, BFFT, BFFT)

julia> expand_dims(Transforms.NoTransform(), Val(2))
(NoTransform, NoTransform)

PencilFFTs.Transforms.is_inplace — Function

is_inplace(transform::AbstractTransform)         -> Bool
is_inplace(transforms::Vararg{AbtractTransform}) -> Union{Bool, Nothing}

Check whether a transform or a list of transforms is performed in-place.

If the list of transforms has a combination of in-place and out-of-place transforms, nothing is returned.

Example

julia> is_inplace(Transforms.RFFT())
false

julia> is_inplace(Transforms.NoTransform!())
true

julia> is_inplace(Transforms.FFT!(), Transforms.R2R!{FFTW.REDFT01}())
true

julia> is_inplace(Transforms.FFT(), Transforms.R2R{FFTW.REDFT01}())
false

julia> is_inplace(Transforms.FFT(), Transforms.R2R!{FFTW.REDFT01}()) === nothing
true

PencilFFTs.Transforms.kind — Function

kind(transform::R2R)

Get kind of real-to-real transform.

PencilFFTs.Transforms.length_output — Function

length_output(transform::AbstractTransform, length_in::Integer)

Returns the length of the transform output, given the length of its input.

The input and output lengths are specified in terms of the respective input and output datatypes. For instance, for real-to-complex transforms, these are respectively the length of input real data and of output complex data.

Also note that for backward real-to-complex transforms (BRFFT), it is assumed that the real data length is even. See also the AbstractFFTs.irfft docs.

Index

PencilFFTs.Transforms

PencilFFTs.Transforms.AbstractCustomPlan

PencilFFTs.Transforms.AbstractTransform

PencilFFTs.Transforms.BFFT

PencilFFTs.Transforms.BFFT!

PencilFFTs.Transforms.BRFFT

PencilFFTs.Transforms.FFT

PencilFFTs.Transforms.FFT!

PencilFFTs.Transforms.IdentityPlan

PencilFFTs.Transforms.IdentityPlan!

PencilFFTs.Transforms.NoTransform

PencilFFTs.Transforms.NoTransform!

PencilFFTs.Transforms.R2R

PencilFFTs.Transforms.R2R!

PencilFFTs.Transforms.RFFT

PencilFFTs.Transforms.binv

PencilFFTs.Transforms.eltype_input

PencilFFTs.Transforms.eltype_output

PencilFFTs.Transforms.expand_dims

PencilFFTs.Transforms.is_inplace

PencilFFTs.Transforms.kind

PencilFFTs.Transforms.length_output

PencilFFTs.Transforms.plan

PencilFFTs.Transforms.scale_factor